refactor(assets): split types.ts into types/ directory module
