+++ assets/js/types/config.ts
@@ -1,4 +1,4 @@
-import type { FixItEventMap } from './core/event-bus'
+import type { MermaidConfig } from './third-party'
 
 /** Mask overlay handler */
 export interface MaskOverlayHandler {
@@ -147,55 +147,6 @@ export interface MathJaxConfig {
   options?: Record<string, any>
 }
 
-export interface MermaidConfig {
-  wrapper?: boolean
-  cdn?: string
-  zenuml?: string
-  themes?: string[]
-  securitylevel?: string
-  look?: string
-  fontfamily?: string
-  layoutloaders?: string[]
-  layout?: string
-}
-
-export interface MermaidRenderResult {
-  svg?: string
-  bindFunctions?: (element: Element) => void
-}
-
-export interface MermaidRuntimeModule {
-  startOnLoad: boolean
-  initialize: (config: Record<string, unknown>) => void
-  render: (id: string, source: string) => Promise<MermaidRenderResult>
-  registerExternalDiagrams?: (diagrams: unknown[]) => Promise<void>
-  registerLayoutLoaders?: (loaders: unknown[]) => void
-}
-
-export interface MermaidRuntime {
-  mermaid: MermaidRuntimeModule
-  config: MermaidConfig
-  zenuml?: unknown
-  loaders: unknown[]
-}
-
-export interface PanzoomTransform {
-  x: number
-  y: number
-  scale: number
-}
-
-export interface PanzoomInstance {
-  getPan: () => { x: number, y: number }
-  getScale: () => number
-  pan: (x: number, y: number, options?: { animate?: boolean, force?: boolean }) => void
-  zoom: (scale: number, options?: { animate?: boolean, force?: boolean }) => void
-  zoomIn: (options?: { animate?: boolean }) => void
-  zoomOut: (options?: { animate?: boolean }) => void
-  zoomWithWheel: (event: WheelEvent) => void
-  reset: (options?: { animate?: boolean }) => void
-}
-
 export interface EncryptionConfig {
   all?: boolean
   shortcode?: boolean
@@ -206,73 +157,4 @@ export interface PrintConfig {
   expandCode?: boolean
   expandDetails?: boolean
   expandFileTree?: boolean
-}
-
-export interface TabContainerChangedDetail {
-  relatedTarget?: Element | null
-}
-
-export type TabContainerChangedEvent = CustomEvent<TabContainerChangedDetail> & {
-  panel: Element | null
-}
-
-type FixItDocumentEventMap = {
-  [K in keyof FixItEventMap]: CustomEvent<FixItEventMap[K]>
-}
-
-/** Public API exposed on window.fixit for backward compatibility. */
-export interface FixItPublicAPI {
-  readonly config: FixItConfig
-  readonly themeMode: string
-  readonly isDark: boolean
-  [method: string]: any
-}
-
-declare global {
-  interface DocumentEventMap extends FixItDocumentEventMap {
-    'tab-container-changed': TabContainerChangedEvent
-  }
-
-  interface Window {
-    // Third-party libraries
-    autocomplete?: any
-    algoliasearch?: any
-    Artalk?: any
-    APlayer?: any
-    CellTooltip?: any
-    cookieconsent?: any
-    CryptoJS?: any
-    echarts?: any
-    Fuse?: any
-    Gitalk?: any
-    JsonViewerElement?: any
-    lgThumbnail?: any
-    lgZoom?: any
-    lightGallery?: any
-    mapboxgl?: any
-    MapboxLanguage?: any
-    MathJax?: any
-    objectFitImages?: () => void
-    pangu?: any
-    postChatUser?: any
-    postChatConfig?: any
-    postChat_theme?: string
-    twemoji?: any
-    twikoo?: any
-    TypeIt?: any
-    Valine?: any
-    Waline?: any
-    Watermark?: any
-    xxhash?: any
-    Panzoom?: (element: SVGElement, options?: Record<string, unknown>) => PanzoomInstance
-
-    // FixIt theme
-    fixit: FixItPublicAPI
-    config: FixItConfig
-    mermaid?: MermaidRuntimeModule
-    _fuseIndex?: any
-    _searchMobile?: any
-    _searchDesktop?: any
-    FixItDecryptor?: any
-  }
 }
