+++ assets/js/types/index.ts
@@ -1,0 +1,3 @@
+export type * from './config'
+export type * from './third-party'
+export type * from './ui'
